+++ crates/goose-server/src/routes/mod.rs
@@ -6,8 +6,10 @@ pub mod extension;
 pub mod health;
 pub mod reply;
 pub mod session;
+pub mod utils;
 
 use axum::Router;
+use crate::routes::reply::routes;
 
 // Function to configure all routes
 pub fn configure(state: crate::state::AppState) -> Router {
